Jack - refactor bad code in home.jsx
