+++ public/css/theme.css
@@ -66,130 +66,7 @@ button:hover {
   box-shadow: inset 0 0 6px rgba(0, 0, 0, 0.3);
 }
 
-/* general box area styling */
 
-.grid {
-  display: grid;
-  margin: auto;
-  max-width: 1520px;
-  grid-template-columns: repeat(3, 1fr);
-  grid-auto-rows: repeat(3, 1fr);
-  grid-template-areas:
-    'global asia europe'
-    'namerica camerica samerica'
-    'africa oceania meast';
-}
-
-.box {
-  font-size: 24px;
-  height: 280px;
-  padding: 30px;
-  margin: 15px;
-  background-color: white;
-  border-radius: 20px;
-  transition: 0.5s;
-  overflow-y: scroll;
-}
-
-.box a {
-  color: inherit;
-  text-decoration: none;
-}
-
-.box a:hover {
-  transition: 0.4s;
-  margin-left: 10px;
-}
-
-/* alphabetized continent boxes */
-
-.africa {
-  color: #d7986b;
-  background-color: #ffcfac;
-  grid-area: africa;
-}
-
-.africa:hover {
-  box-shadow: 10px 10px 0px #d7986b;
-}
-
-.asia {
-  color: #deba04;
-  background-color: #f8e16b;
-  grid-area: asia;
-}
-
-.asia:hover {
-  box-shadow: 10px 10px 0px #c4a400;
-}
-
-.centralamerica {
-  color: #e78b36;
-  background-color: #ffbb7c;
-}
-
-.centralamerica:hover {
-  box-shadow: 10px 10px 0px #e78b36;
-}
-
-.europe {
-  color: #69b42f;
-  background-color: #bbe19e;
-  grid-area: europe;
-}
-
-.europe:hover {
-  box-shadow: 10px 10px 0px #69b42f;
-}
-
-.global {
-  color: #b1bf05;
-  background-color: #eaf473;
-  grid-area: global;
-}
-
-.global:hover {
-  box-shadow: 10px 10px 0px #b1bf05;
-}
-
-.middleeast {
-  color: #9843ac;
-  background-color: #c592d1;
-  grid-area: meast;
-}
-
-.middleeast:hover {
-  box-shadow: 10px 10px 0px #9843ac;
-}
-
-.northamerica {
-  color: #cc585e;
-  background-color: #f4a0a5;
-}
-
-.northamerica:hover {
-  box-shadow: 10px 10px 0px #cc585e;
-}
-
-.oceania {
-  color: #5271c1;
-  background-color: #9eb1e1;
-  grid-area: oceania;
-}
-
-.oceania:hover {
-  box-shadow: 10px 10px 0px #5271c1;
-}
-
-.southamerica {
-  color: #4598b3;
-  background-color: #9ed1e1;
-  grid-area: namerica;
-}
-
-.southamerica:hover {
-  box-shadow: 10px 10px 0px #4598b3;
-}
 
 /* footer */
