+++ src/components/home_page/Home.jsx
@@ -1,0 +1,484 @@
+import React from 'react';
+import './Home.css'
+
+
+const GLOBAL = [
+{name:"Children's Hunger Fund", url:'https://bit.ly/donateCHFcovid19'},
+{name:"Unicef'", url:'https://uni.cf/2V8fODA'},
+{name:"United Nations", url:'https://bit.ly/donateUNcovid19'},
+{name:"World Health Organization", url:'http://bit.ly/donateWHO'}
+]
+
+
+// TODO: get rid of repetitive HTML and take that data and put it into objects
+
+const ASIA = [
+  {name:"Children's Hunger Fund", url:'https://bit.ly/donateCHFcovid19'},
+  {name:"Unicef'", url:'https://uni.cf/2V8fODA'},
+  {name:"United Nations", url:'https://bit.ly/donateUNcovid19'},
+  {name:"World Health Organization", url:'http://bit.ly/donateWHO'}
+]
+
+const Link = (props) => {
+  // props: String url, String name
+   return (
+    <p>
+    <a
+      href={props.url}
+      rel='noopener noreferrer'
+      target='_blank'
+    >
+      {props.name}
+    </a>
+  </p>
+   )
+}
+
+const Box = (props) => {
+  // props: String name, List<Link> list_of_links, String className
+  return (
+        <div className={`box ${props.className}`} id='box'>
+          <h1> {props.name} </h1>
+          {
+            props.list_of_links.map((link) => {
+              return (
+                <Link url = {link.url} name = {link.name} />
+              )
+            })
+          }
+      </div>
+  )
+}
+
+
+function Home() {
+  return (
+    <div className='home'>
+      <title>Donate to Fight COVID-19</title>
+      <div class='grid'>
+        <Box name = "Global" list_of_links = {GLOBAL} className = "global" />
+
+        <div class='box asia' id='box'>
+          <h1> asia </h1>
+          <p>
+            <a
+              href='http://bit.ly/donateAFcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Afghanistan
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateCNcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              China
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateHKcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Hong Kong (Red Cross)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateSTCcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Hong Kong (Save the Children)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateINDcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              India (Rapid Response)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateGIcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              India (Give)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateINcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Indonesia
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateMYcovid2019'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Malaysia
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donatePKcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Pakistan
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donatePHcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              {' '}
+              Philippines
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donatePHIcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Philippines (Masterlist)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateSGcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Singapore
+            </a>
+          </p>
+        </div>
+
+        <div class='box europe' id='box'>
+          <h1> europe </h1>
+          <p>
+            <a
+              href='https://bit.ly/donateDEcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Germany
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateITcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Italy
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateNLcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Netherlands
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donatePLcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Poland
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donatePTGcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Portugal (Go Parity)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donatePTcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Portugal (Red Cross)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateSIcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Slovenia
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateEScovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Spain
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateUKcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              UK
+            </a>
+          </p>
+        </div>
+
+        <div class='box northamerica' id='box'>
+          <h1> north america </h1>
+          <p>
+            {' '}
+            <strong> USA: </strong>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateEMcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Atlanta, GA (Emory)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateATLcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Atlanta, GA (Food Bank)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateATXcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Austin, TX
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateCILcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Chicago, IL
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateHTXcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Houston, TX
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateNILcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Illinois
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateLAcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Los Angeles, CA
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateNYcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              New York City, NY
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donatePITcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Pittsburgh, PA
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateWAcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Seattle, WA (University of Washington)
+            </a>
+          </p>
+          <p>
+            <a
+              href='http://bit.ly/donateUWcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Seattle, WA (Seattle Foundation)
+            </a>
+          </p>
+          <p>
+            {' '}
+            <strong> Canada: </strong>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateCNDcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Canada (Red Cross)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateFVHcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Canada (Fraser Valley)
+            </a>
+          </p>
+        </div>
+
+        <div class='box centralamerica' id='box'>
+          <h1> central america </h1>
+          <p>
+            <a
+              href='https://bit.ly/donateARcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Aruba
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateDOcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Dominican Republic
+            </a>
+          </p>
+        </div>
+
+        <div class='box southamerica' id='box'>
+          <h1> south america </h1>
+          <p>
+            <a
+              href='https://bit.ly/donateBRcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Brazil (BSocial)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateFIcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Brazil (Fiocruz)
+            </a>
+          </p>
+          <p>
+            <a
+              href='https://bit.ly/donateUFcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Brazil (UFRGS)
+            </a>
+          </p>
+        </div>
+
+        <div class='box africa' id='box'>
+          <h1> africa </h1>
+          <p>
+            <a
+              href='http://bit.ly/donateUGcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Uganda
+            </a>
+          </p>
+        </div>
+
+        <div class='box oceania' id='box'>
+          <h1> oceania </h1>
+          <p>
+            <a
+              href='http://bit.ly/donateAUScovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              Australia
+            </a>
+          </p>
+        </div>
+
+        <div class='box middleeast' id='box'>
+          <h1> middle east </h1>
+          <p>
+            <a
+              href='http://bit.ly/donateAEcovid19'
+              rel='noopener noreferrer'
+              target='_blank'
+            >
+              UAE
+            </a>
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default Home;
