+++ src/components/index.js
@@ -1,5 +1,5 @@
 export { default as Navigation } from "./Navigation";
-export { default as Home } from "./Home";
+export { default as Home } from "./home_page/Home.jsx";
 export { default as Footer } from "./Footer";
 export { default as Form } from "./Form";
 export { default as Ideas} from "./Ideas";
